[FIX] change name of new objects
