+++ sale_credit_note/wizard/new_credit_note_view.xml
@@ -10,7 +10,7 @@
     <data>
         <record id="new_credit_note_view" model="ir.ui.view">
             <field name="name">New Credit Note</field>
-            <field name="model">new.credit.note</field>
+            <field name="model">sale.new.credit.note</field>
             <field name="arch" type="xml">
                 <form string="New credit note" version="7.0">
                     <p class="oe_grey">
@@ -32,7 +32,7 @@
          <record id="action_credit_note" model="ir.actions.act_window">
             <field name="name">New credit note</field>
             <field name="type">ir.actions.act_window</field>
-            <field name="res_model">new.credit.note</field>
+            <field name="res_model">sale.new.credit.note</field>
             <field name="view_type">form</field>
             <field name="view_mode">form</field>
             <field name="target">new</field>
